src/requirements.txt
matplotlib
numpy
pandas
loralib
Pillow
scikit-image
opencv-python
tqdm
wandb
transformers
timm
peft
huggingface_hub